requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
flask==3.0.0
google-cloud-translate==3.14.0
google-generativeai==0.3.0
